Improved error code compatibility
libusb-1.0 can return result codes 0, -1 through -12 and -99.  The old
libusb-0.1 software was always returning the negative of an errno.h
error code.

Unfortunately the current version of libusb-compat passes through some
of the libusb-1.0 error codes unchanged.  (Some of the error codes are
translated for some functions and different ones for other functions
but plenty are not translated.)

The patch attached to the below bug report makes sure all libusb-1.0
error codes are translated into the libusb-0.1 equivalent negative of
an errno.h define value.  Additionally, this patch makes sure that
errno is always being set as some callers of libusb-0.1 expect that to
be the case.  Finally since errno is now always being set by libusb-
compat, the implementation of usb_strerror is changed to return
strerror(errno) instead of the unhelpful "Unknown error".